requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
python-dotenv==1.0.1
loguru==0.7.2

Uvicorn==0.30.6
asgiref==3.8.1

